requirements.txt
Django==5.1.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
django-environ==0.11.2
django-debug-toolbar==4.4.6
djangorestframework-simplejwt==5.3.1
